+++ packages/dataviews/src/dataform-layouts/panel/dropdown.tsx
@@ -12,6 +12,7 @@ import {
 import { __ } from '@wordpress/i18n';
 import { useMemo } from '@wordpress/element';
 import { closeSmall } from '@wordpress/icons';
+import { useFocusOnMount } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -115,11 +116,13 @@ function PanelDropdown< Item >( {
 		[ popoverAnchor ]
 	);
 
+	const focusOnMountRef = useFocusOnMount( 'firstInputElement' );
+
 	return (
 		<Dropdown
 			contentClassName="dataforms-layouts-panel__field-dropdown"
 			popoverProps={ popoverProps }
-			focusOnMount
+			focusOnMount={ false }
 			toggleProps={ {
 				size: 'compact',
 				variant: 'tertiary',
@@ -139,25 +142,31 @@ function PanelDropdown< Item >( {
 			renderContent={ ( { onClose } ) => (
 				<>
 					<DropdownHeader title={ fieldLabel } onClose={ onClose } />
-					<DataFormLayout
-						data={ data }
-						form={ form }
-						onChange={ onChange }
-						validity={ formValidity }
-					>
-						{ ( FieldLayout, childField, childFieldValidity ) => (
-							<FieldLayout
-								key={ childField.id }
-								data={ data }
-								field={ childField }
-								onChange={ onChange }
-								hideLabelFromVision={
-									( form?.fields ?? [] ).length < 2
-								}
-								validity={ childFieldValidity }
-							/>
-						) }
-					</DataFormLayout>
+					<div ref={ focusOnMountRef }>
+						<DataFormLayout
+							data={ data }
+							form={ form }
+							onChange={ onChange }
+							validity={ formValidity }
+						>
+							{ (
+								FieldLayout,
+								childField,
+								childFieldValidity
+							) => (
+								<FieldLayout
+									key={ childField.id }
+									data={ data }
+									field={ childField }
+									onChange={ onChange }
+									hideLabelFromVision={
+										( form?.fields ?? [] ).length < 2
+									}
+									validity={ childFieldValidity }
+								/>
+							) }
+						</DataFormLayout>
+					</div>
 				</>
 			) }
 		/>